+++ src/example.ts
@@ -1,61 +1,66 @@
-import dwani from './index';
+import dwani from 'dwani';
 import * as fs from 'fs/promises';
+import { ChatRequest, VisionRequest, ASRRequest, TTSRequest, TranslateRequest, DocumentsRequest } from 'dwani';
 
 async function main() {
   try {
-    // Chat example
-    const chatResponse = await dwani.Chat.create({
-      prompt: 'Hello!',
+    // Chat example: Multilingual conversation
+    const chatParams: ChatRequest = {
+      prompt: 'What is the capital of France?',
       src_lang: 'eng_Latn',
       tgt_lang: 'kan_Knda',
-    });
-    console.log('Chat:', chatResponse);
+    };
+    const chatResponse = await dwani.Chat.create(chatParams);
+    console.log('Chat Response:', JSON.stringify(chatResponse, null, 2));
 
-    /*
-    // Vision example
-    const visionResponse = await dwani.Vision.caption({
-      file_path: './image.png',
-      query: 'Describe this image',
+    // Vision example: Image captioning
+    const visionParams: VisionRequest = {
+      file_path: './sample_image.jpg',
+      query: 'Describe the contents of this image in detail.',
       src_lang: 'eng_Latn',
       tgt_lang: 'deu_Latn',
-    });
-    console.log('Vision:', visionResponse);
+    };
+    const visionResponse = await dwani.Vision.caption(visionParams);
+    console.log('Vision Response:', JSON.stringify(visionResponse, null, 2));
 
-    // ASR example
-    const asrResponse = await dwani.ASR.transcribe({
-      file_path: './audio.wav',
+    // ASR example: Audio transcription
+    const asrParams: ASRRequest = {
+      file_path: './sample_audio.wav',
       language: 'french',
-    });
-    console.log('ASR:', asrResponse);
+    };
+    const asrResponse = await dwani.ASR.transcribe(asrParams);
+    console.log('ASR Response:', JSON.stringify(asrResponse, null, 2));
 
-    // TTS example
-    const ttsResponse = await dwani.Audio.speech({
-      input: 'Bonjour le monde',
+    // TTS example: Text-to-speech conversion
+    const ttsParams: TTSRequest = {
+      input: 'Bonjour, comment allez-vous?',
       response_format: 'mp3',
-    });
-    await fs.writeFile('output.mp3', ttsResponse);
-    console.log('TTS: Audio saved to output.mp3');
+    };
+    const ttsResponse = await dwani.Audio.speech(ttsParams);
+    await fs.writeFile('output_audio.mp3', ttsResponse);
+    console.log('TTS: Audio saved to output_audio.mp3');
 
-    // Translation example
-    const translateResponse = await dwani.Translate.run_translate({
-      sentences: ['Hello'],
+    // Translation example: Sentence translation
+    const translateParams: TranslateRequest = {
+      sentences: ['The weather is nice today.', 'I am learning TypeScript.'],
       src_lang: 'eng_Latn',
       tgt_lang: 'spa_Latn',
-    });
-    console.log('Translate:', translateResponse);
-*/
-    // Documents example
-    const docResponse = await dwani.Documents.run_extract({
-      file_path: './document.pdf',
+    };
+    const translateResponse = await dwani.Translate.run_translate(translateParams);
+    console.log('Translation Response:', JSON.stringify(translateResponse, null, 2));
+
+    // Documents example: PDF text extraction
+    const docParams: DocumentsRequest = {
+      file_path: './sample_document.pdf',
       page_number: 1,
       src_lang: 'eng_Latn',
       tgt_lang: 'ita_Latn',
-    });
-    console.log('Documents:', docResponse);
+    };
+    const docResponse = await dwani.Documents.run_extract(docParams);
+    console.log('Documents Response:', JSON.stringify(docResponse, null, 2));
   } catch (error: any) {
     console.error('Error:', error.message);
   }
-    
 }
 
 main();
