--- src/examples/typescript_example.d.ts
@@ -1,1 +1,0 @@
-export {};
