+++ tsconfig.json
@@ -9,5 +9,5 @@
     "esModuleInterop": true,
     "skipLibCheck": true
   },
-  "exclude": ["src/examples"]
+"exclude": ["src/samples"]
 }
